refactor(layout): split xml_parser into three focused modules
The 800-line xml_parser.py mixed three responsibilities. Split it along the
seam established by the earlier refactors:

- layout_core.py  -- format-free graph layout + routing (_sugiyama_layout,
  _route_edges, _loop_waypoints)
- bpmn_writer.py  -- semantic BPMN + BPMNDI (_build_semantic_process,
  _add_diagram, json_to_bpmn)
- pnml_writer.py  -- PNML layout + client cleanup (assign_pnml_coordinates,
  _clean_pnml, _set_arc_waypoints, _label_width)

Pure move (functions unchanged), plus import updates in bpmn_builder, routes
and the test. Behaviour preserved (106 tests green).
