+++ app/backend/bpmn_writer.py
@@ -1,0 +1,199 @@
+"""Build BPMN 2.0 XML (semantic process + diagram interchange) from a model."""
+
+import logging
+import xml.etree.ElementTree as ET
+
+from app.backend.layout_core import _route_edges, _sugiyama_layout
+
+logger = logging.getLogger(__name__)
+
+_BPMN_EVENT_W, _BPMN_EVENT_H = 36, 36
+_BPMN_TASK_W, _BPMN_TASK_H = 100, 80
+_BPMN_GATEWAY_W, _BPMN_GATEWAY_H = 50, 50
+
+# Maps a model event "type" to its BPMN element tag; anything unrecognised
+# becomes a generic intermediateCatchEvent.
+_EVENT_TYPE_MAP = {
+    "Start": "startEvent",
+    "startEvent": "startEvent",
+    "End": "endEvent",
+    "endEvent": "endEvent",
+}
+
+# BPMN namespaces, registered once so ET.tostring emits the expected prefixes.
+_NS = {
+    "bpmn": "http://www.omg.org/spec/BPMN/20100524/MODEL",
+    "bpmndi": "http://www.omg.org/spec/BPMN/20100524/DI",
+    "di": "http://www.omg.org/spec/DD/20100524/DI",
+    "dc": "http://www.omg.org/spec/DD/20100524/DC",
+    "xsi": "http://www.w3.org/2001/XMLSchema-instance",
+}
+ET.register_namespace("", _NS["bpmn"])
+ET.register_namespace("bpmndi", _NS["bpmndi"])
+ET.register_namespace("di", _NS["di"])
+ET.register_namespace("dc", _NS["dc"])
+ET.register_namespace("xsi", _NS["xsi"])
+
+
+def _build_semantic_process(model):
+    """Build the semantic BPMN ``<definitions>``/``<process>`` tree.
+
+    Translates each element's type into its BPMN tag and adds every event,
+    task, gateway and sequence flow. No diagram geometry is produced here.
+    """
+    definitions = ET.Element(
+        f"{{{_NS['bpmn']}}}definitions",
+        attrib={
+            f"{{{_NS['xsi']}}}schemaLocation": "http://www.omg.org/spec/BPMN/20100524/MODEL https://www.omg.org/spec/BPMN/20100501/BPMN20.xsd",
+            "targetNamespace": "http://example.bpmn.com/schema/bpmn",
+        },
+    )
+    process = ET.SubElement(
+        definitions,
+        f"{{{_NS['bpmn']}}}process",
+        attrib={"id": "Process_1", "isExecutable": "false"},
+    )
+
+    semantic_elements = {}
+
+    for event in model["events"]:
+        event_type = _EVENT_TYPE_MAP.get(event["type"], "intermediateCatchEvent")
+        semantic_elements[event["id"]] = ET.SubElement(
+            process,
+            f"{{{_NS['bpmn']}}}{event_type}",
+            id=event["id"],
+            name=event["name"],
+        )
+
+    for task in model["tasks"]:
+        # Convert task type to camelCase (e.g., UserTask -> userTask).
+        task_type = task["type"][0].lower() + task["type"][1:]
+        semantic_elements[task["id"]] = ET.SubElement(
+            process, f"{{{_NS['bpmn']}}}{task_type}", id=task["id"], name=task["name"]
+        )
+
+    for gateway in model["gateways"]:
+        gateway_type = gateway["type"][0].lower() + gateway["type"][1:]
+        semantic_elements[gateway["id"]] = ET.SubElement(
+            process,
+            f"{{{_NS['bpmn']}}}{gateway_type}",
+            id=gateway["id"],
+            name=gateway["name"],
+        )
+
+    for flow in model["flows"]:
+        source = semantic_elements[flow["source"]]
+        target = semantic_elements[flow["target"]]
+        # These per-node <incoming>/<outgoing> tags look redundant with the
+        # sequenceFlow's source/target below, but the model-transformer needs
+        # them: its get_in_degree/get_out_degree read the node tags, not the
+        # edges. Without them it does not error -- it silently builds a wrong
+        # net (gateways look degree-0 and get pruned). Do not remove.
+        ET.SubElement(source, f"{{{_NS['bpmn']}}}outgoing").text = flow["id"]
+        ET.SubElement(target, f"{{{_NS['bpmn']}}}incoming").text = flow["id"]
+        ET.SubElement(
+            process,
+            f"{{{_NS['bpmn']}}}sequenceFlow",
+            id=flow["id"],
+            sourceRef=flow["source"],
+            targetRef=flow["target"],
+        )
+
+    return definitions
+
+
+def _add_diagram(definitions, model):
+    """Add the BPMN diagram interchange (layout + shapes + edges).
+
+    Sizes every node, computes a layered layout, then emits a BPMNShape per node
+    and a BPMNEdge per flow. Edges are routed orthogonally: vertical segments
+    ("risers") live in the gaps between columns, never inside one, and each riser
+    in a gap gets its own channel so arrows do not run on top of one another.
+    Edges that skip layers are threaded through reserved dummy lanes (see
+    :func:`_sugiyama_layout`) so they do not cross the nodes in between.
+    ``verify`` upstream guarantees every node and flow endpoint exists, so
+    positions are looked up directly.
+    """
+    bpmn_di = ET.SubElement(
+        definitions, f"{{{_NS['bpmndi']}}}BPMNDiagram", attrib={"id": "BPMNDiagram_1"}
+    )
+    bpmn_plane = ET.SubElement(
+        bpmn_di,
+        f"{{{_NS['bpmndi']}}}BPMNPlane",
+        attrib={"id": "BPMNPlane_1", "bpmnElement": "Process_1"},
+    )
+
+    sizes: dict[str, dict] = {}
+    for event in model["events"]:
+        sizes[event["id"]] = {"w": _BPMN_EVENT_W, "h": _BPMN_EVENT_H}
+    for task in model["tasks"]:
+        sizes[task["id"]] = {"w": _BPMN_TASK_W, "h": _BPMN_TASK_H}
+    for gateway in model["gateways"]:
+        sizes[gateway["id"]] = {"w": _BPMN_GATEWAY_W, "h": _BPMN_GATEWAY_H}
+
+    positions, ctx = _sugiyama_layout(
+        sizes, model["flows"], h_gap=180, v_gap=90, x_offset=50, y_offset=50
+    )
+
+    for element in model["tasks"] + model["events"] + model["gateways"]:
+        bpmn_shape = ET.SubElement(
+            bpmn_plane,
+            f"{{{_NS['bpmndi']}}}BPMNShape",
+            attrib={"id": f"{element['id']}_di", "bpmnElement": element["id"]},
+        )
+        pos = positions[element["id"]]
+        ET.SubElement(
+            bpmn_shape,
+            f"{{{_NS['dc']}}}Bounds",
+            attrib={
+                "x": str(pos["x"]),
+                "y": str(pos["y"]),
+                "width": str(pos["w"]),
+                "height": str(pos["h"]),
+            },
+        )
+
+    if not model["flows"]:
+        return
+
+    routes = _route_edges(positions, ctx, model["flows"], "full_ortho")
+    for idx, flow in enumerate(model["flows"]):
+        bpmn_edge = ET.SubElement(
+            bpmn_plane,
+            f"{{{_NS['bpmndi']}}}BPMNEdge",
+            attrib={"id": f"{flow['id']}_di", "bpmnElement": flow["id"]},
+        )
+        prev = None
+        for px, py in routes[idx]:
+            point = (str(int(round(px))), str(int(round(py))))
+            if point == prev:
+                continue
+            ET.SubElement(
+                bpmn_edge,
+                f"{{{_NS['di']}}}waypoint",
+                attrib={"x": point[0], "y": point[1]},
+            )
+            prev = point
+
+
+def json_to_bpmn(model, include_layout=True):
+    """Convert a validated logical process model into BPMN 2.0 XML.
+
+    Builds the semantic process and serializes it. When ``include_layout`` is
+    true it also draws the diagram (shapes and waypoints). The PNML path passes
+    ``include_layout=False``: the transformer ignores BPMN layout and we lay the
+    PNML out separately, so computing a BPMN layout there would be wasted work.
+    """
+    logger.info(
+        "Converting model to BPMN",
+        extra={k: len(model[k]) for k in ("events", "tasks", "gateways", "flows")},
+    )
+    definitions = _build_semantic_process(model)
+    if include_layout:
+        _add_diagram(definitions, model)
+
+    tree = ET.ElementTree(definitions)
+    ET.indent(tree, space="  ", level=0)
+    return ET.tostring(definitions, encoding="utf-8", xml_declaration=True).decode(
+        "utf-8"
+    )
