+++ app/backend/layout_core.py
@@ -1,48 +1,12 @@
-import xml.etree.ElementTree as ET
-import logging
+"""Layered (Sugiyama) layout + orthogonal edge routing, independent of XML format.
+
+Pure graph math: nodes are ``{id: {"w", "h"}}`` and edges
+``[{"source", "target"}]``; the output is node positions plus waypoint polylines.
+The BPMN and PNML writers feed this and turn the result into their own diagram
+vocabulary.
+"""
+
 from collections import deque, defaultdict
-
-logger = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Layout constants
-# ---------------------------------------------------------------------------
-_BPMN_EVENT_W, _BPMN_EVENT_H = 36, 36
-_BPMN_TASK_W, _BPMN_TASK_H = 100, 80
-_BPMN_GATEWAY_W, _BPMN_GATEWAY_H = 50, 50
-
-_PNML_PLACE_W, _PNML_PLACE_H = 50, 50
-_PNML_TRANS_W, _PNML_TRANS_H = 50, 30
-
-# Approximate glyph width (px) of WoPeD's small label font (~11px). Labels are
-# drawn centred BELOW the node, so the layout must reserve at least the label's
-# width per column or adjacent labels overlap. Used to widen columns to fit the
-# text instead of the node box alone.
-_LABEL_CHAR_PX = 7
-
-# Maps a model event "type" to its BPMN element tag; anything unrecognised
-# becomes a generic intermediateCatchEvent.
-_EVENT_TYPE_MAP = {
-    "Start": "startEvent",
-    "startEvent": "startEvent",
-    "End": "endEvent",
-    "endEvent": "endEvent",
-}
-
-# BPMN namespaces, registered once so ET.tostring emits the expected prefixes.
-_NS = {
-    "bpmn": "http://www.omg.org/spec/BPMN/20100524/MODEL",
-    "bpmndi": "http://www.omg.org/spec/BPMN/20100524/DI",
-    "di": "http://www.omg.org/spec/DD/20100524/DI",
-    "dc": "http://www.omg.org/spec/DD/20100524/DC",
-    "xsi": "http://www.w3.org/2001/XMLSchema-instance",
-}
-ET.register_namespace("", _NS["bpmn"])
-ET.register_namespace("bpmndi", _NS["bpmndi"])
-ET.register_namespace("di", _NS["di"])
-ET.register_namespace("dc", _NS["dc"])
-ET.register_namespace("xsi", _NS["xsi"])
 
 
 def _sugiyama_layout(
@@ -273,266 +237,6 @@ def _sugiyama_layout(
     return positions, ctx
 
 
-def _set_arc_waypoints(arc_el, points, ns_prefix):
-    """Write *points* as intermediate ``<graphics><position>`` bend points on an
-    ``<arc>``.
-
-    WoPeD reads each ``<position>`` as an absolute-canvas bend point between the
-    arc's (implicit) source and target anchors, so only the interior points are
-    written -- the endpoints are derived from the node positions by the client.
-    """
-    graphics = arc_el.find(f"{ns_prefix}graphics")
-    if graphics is None:
-        graphics = ET.SubElement(arc_el, f"{ns_prefix}graphics")
-    for px, py in points:
-        ET.SubElement(
-            graphics,
-            f"{ns_prefix}position",
-            attrib={"x": str(int(round(px))), "y": str(int(round(py)))},
-        )
-
-
-def _label_width(elem, ns_prefix):
-    """Estimate the rendered width (px) of a node's label from ``<name><text>``.
-
-    WoPeD draws the name centred below the node, so a long name needs a wider
-    column than the node box to keep it from overlapping its neighbours.
-    Returns 0 for unnamed nodes (e.g. silent places).
-    """
-    name = elem.find(f"{ns_prefix}name")
-    text = name.find(f"{ns_prefix}text") if name is not None else None
-    if text is None or not text.text:
-        return 0
-    return len(text.text.strip()) * _LABEL_CHAR_PX
-
-
-def assign_pnml_coordinates(pnml_xml):
-    """Parse a PNML XML string and assign proper layout coordinates to all
-    places and transitions.
-
-    Uses the same layered layout algorithm as the BPMN generator.  In PNML,
-    ``<graphics><position>`` holds *centre* coordinates, so positions returned
-    by ``_sugiyama_layout`` (top-left) are shifted by half the element size.
-
-    Back-edges (rework loops) are routed the same way as in BPMN: each gets its
-    own horizontal lane beneath the diagram, written as ``<arc>`` bend points,
-    so loops do not depend on each client's own arc routing (which differs and
-    would otherwise cut a straight line back across the columns).
-
-    If *pnml_xml* is not valid XML the original string is returned unchanged.
-
-    Args:
-        pnml_xml: PNML XML string (with or without ``<?xml ...?>`` declaration).
-
-    Returns:
-        Updated PNML XML string.
-    """
-    if not pnml_xml or not isinstance(pnml_xml, str):
-        return pnml_xml
-
-    try:
-        root = ET.fromstring(pnml_xml)
-    except ET.ParseError:
-        logger.warning("assign_pnml_coordinates: not valid XML – layout skipped")
-        return pnml_xml
-
-    # Detect Clark-notation namespace prefix, e.g. '{http://www.pnml.org/...}'.
-    raw_tag = root.tag
-    ns_prefix = (
-        "{" + raw_tag[1 : raw_tag.index("}")] + "}" if raw_tag.startswith("{") else ""
-    )
-
-    # Register the namespace so ET.tostring() preserves the default namespace.
-    if ns_prefix:
-        ET.register_namespace("", ns_prefix[1:-1])
-
-    # Collect all places and transitions from the entire tree (handles
-    # nested <pnml><net><page>... hierarchies).
-    elements_by_id: dict[str, dict] = {}
-    elem_xml_map: dict[str, ET.Element] = {}
-
-    for place in root.iter(f"{ns_prefix}place"):
-        pid = place.get("id")
-        if pid:
-            w = max(_PNML_PLACE_W, _label_width(place, ns_prefix))
-            elements_by_id[pid] = {"w": w, "h": _PNML_PLACE_H}
-            elem_xml_map[pid] = place
-
-    for trans in root.iter(f"{ns_prefix}transition"):
-        tid = trans.get("id")
-        if tid:
-            w = max(_PNML_TRANS_W, _label_width(trans, ns_prefix))
-            elements_by_id[tid] = {"w": w, "h": _PNML_TRANS_H}
-            elem_xml_map[tid] = trans
-
-    if not elements_by_id:
-        return pnml_xml  # nothing to lay out
-
-    arcs = [
-        (arc, arc.get("source"), arc.get("target"))
-        for arc in root.iter(f"{ns_prefix}arc")
-        if arc.get("source") and arc.get("target")
-    ]
-    flows = [{"source": src, "target": tgt} for _, src, tgt in arcs]
-
-    positions, ctx = _sugiyama_layout(
-        elements_by_id, flows, h_gap=80, v_gap=50, x_offset=100, y_offset=100
-    )
-
-    for nid, pos in positions.items():
-        elem = elem_xml_map.get(nid)
-        if elem is None:
-            continue
-
-        # PNML <position> stores centre coordinates.
-        cx = pos["x"] + pos["w"] // 2
-        cy = pos["y"] + pos["h"] // 2
-
-        graphics = elem.find(f"{ns_prefix}graphics")
-        if graphics is None:
-            graphics = ET.SubElement(elem, f"{ns_prefix}graphics")
-
-        position_el = graphics.find(f"{ns_prefix}position")
-        if position_el is None:
-            position_el = ET.SubElement(graphics, f"{ns_prefix}position")
-        position_el.set("x", str(cx))
-        position_el.set("y", str(cy))
-
-    # Route every back-edge (rework loop) through its own lane beneath the
-    # diagram, mirroring the BPMN generator. Without explicit bend points each
-    # client auto-routes the loop differently (a straight line back across the
-    # columns), so the rendering is unstable across the fat client and web.
-    # Forward arcs stay waypoint-free (loops_only); arcs and flows are aligned.
-    routes = _route_edges(positions, ctx, flows, "loops_only")
-    for idx, (arc, _src, _tgt) in enumerate(arcs):
-        points = routes[idx]
-        if not points:
-            continue
-        # Drop the first/last point: those are the node anchors WoPeD derives
-        # itself; only the interior U-bend points are stored as bend points.
-        _set_arc_waypoints(arc, points[1:-1], ns_prefix)
-
-    # Tidy the PNML for clients (empty labels, empty markers, id="" noise).
-    # Independent of layout, so kept out of the layout steps above.
-    _clean_pnml(root, ns_prefix)
-
-    ET.indent(ET.ElementTree(root), space="  ", level=0)
-    return ET.tostring(root, encoding="unicode")
-
-
-def _clean_pnml(root, ns_prefix):
-    """Tidy the transformer's PNML for clients, independent of layout.
-
-    Three client-facing fixes that are not geometry:
-
-    * Anonymous nodes (silent/start/end places, operator helper transitions)
-      carry no ``<name>``, so WoPeD would show the raw id ("SILENTFROMxTOy",
-      ...) as a label. Give them an empty ``<name>`` so they render unlabelled.
-    * The transformer marks every UserTask with a WoPeD
-      ``<trigger>``/``<transitionResource>`` -- that marker is how the reverse
-      (pnml->bpmn) direction tells a UserTask from a plain Task, so it is kept
-      even with no role/orga. An EMPTY marker is meaningless to clients, so it
-      (and any ``<toolspecific>`` left empty by it) is dropped.
-    * pydantic-xml gives every element a default empty ``id=""``; real node/arc
-      ids are never empty, so the noise ids are stripped.
-    """
-    nodes = list(root.iter(f"{ns_prefix}place")) + list(
-        root.iter(f"{ns_prefix}transition")
-    )
-    for elem in nodes:
-        name_el = elem.find(f"{ns_prefix}name")
-        if name_el is None:
-            name_el = ET.Element(f"{ns_prefix}name")
-            ET.SubElement(name_el, f"{ns_prefix}text").text = ""
-            elem.insert(0, name_el)
-
-        for ts in elem.findall(f"{ns_prefix}toolspecific"):
-            trigger = ts.find(f"{ns_prefix}trigger")
-            resource = ts.find(f"{ns_prefix}transitionResource")
-            resource_empty = resource is not None and not (
-                resource.get("roleName") or resource.get("organizationalUnitName")
-            )
-            if trigger is not None and trigger.get("type") == "200" and resource_empty:
-                ts.remove(trigger)
-                ts.remove(resource)
-            if not any(
-                ts.find(f"{ns_prefix}{tag}") is not None
-                for tag in ("operator", "trigger", "transitionResource", "subprocess")
-            ):
-                elem.remove(ts)
-
-    for el in root.iter():
-        if el.get("id") == "":
-            del el.attrib["id"]
-
-
-def _build_semantic_process(model):
-    """Build the semantic BPMN ``<definitions>``/``<process>`` tree.
-
-    Translates each element's type into its BPMN tag and adds every event,
-    task, gateway and sequence flow. No diagram geometry is produced here.
-    """
-    definitions = ET.Element(
-        f"{{{_NS['bpmn']}}}definitions",
-        attrib={
-            f"{{{_NS['xsi']}}}schemaLocation": "http://www.omg.org/spec/BPMN/20100524/MODEL https://www.omg.org/spec/BPMN/20100501/BPMN20.xsd",
-            "targetNamespace": "http://example.bpmn.com/schema/bpmn",
-        },
-    )
-    process = ET.SubElement(
-        definitions,
-        f"{{{_NS['bpmn']}}}process",
-        attrib={"id": "Process_1", "isExecutable": "false"},
-    )
-
-    semantic_elements = {}
-
-    for event in model["events"]:
-        event_type = _EVENT_TYPE_MAP.get(event["type"], "intermediateCatchEvent")
-        semantic_elements[event["id"]] = ET.SubElement(
-            process,
-            f"{{{_NS['bpmn']}}}{event_type}",
-            id=event["id"],
-            name=event["name"],
-        )
-
-    for task in model["tasks"]:
-        # Convert task type to camelCase (e.g., UserTask -> userTask).
-        task_type = task["type"][0].lower() + task["type"][1:]
-        semantic_elements[task["id"]] = ET.SubElement(
-            process, f"{{{_NS['bpmn']}}}{task_type}", id=task["id"], name=task["name"]
-        )
-
-    for gateway in model["gateways"]:
-        gateway_type = gateway["type"][0].lower() + gateway["type"][1:]
-        semantic_elements[gateway["id"]] = ET.SubElement(
-            process,
-            f"{{{_NS['bpmn']}}}{gateway_type}",
-            id=gateway["id"],
-            name=gateway["name"],
-        )
-
-    for flow in model["flows"]:
-        source = semantic_elements[flow["source"]]
-        target = semantic_elements[flow["target"]]
-        # These per-node <incoming>/<outgoing> tags look redundant with the
-        # sequenceFlow's source/target below, but the model-transformer needs
-        # them: its get_in_degree/get_out_degree read the node tags, not the
-        # edges. Without them it does not error -- it silently builds a wrong
-        # net (gateways look degree-0 and get pruned). Do not remove.
-        ET.SubElement(source, f"{{{_NS['bpmn']}}}outgoing").text = flow["id"]
-        ET.SubElement(target, f"{{{_NS['bpmn']}}}incoming").text = flow["id"]
-        ET.SubElement(
-            process,
-            f"{{{_NS['bpmn']}}}sequenceFlow",
-            id=flow["id"],
-            sourceRef=flow["source"],
-            targetRef=flow["target"],
-        )
-
-    return definitions
-
-
 def _loop_waypoints(src, tgt, lane_y):
     """Route a back-edge / same-layer edge as a U through the *lane_y* corridor.
 
@@ -707,100 +411,3 @@ def _route_edges(positions, ctx, flows, strategy):
         else:
             routes[idx] = []
     return routes
-
-
-def _add_diagram(definitions, model):
-    """Add the BPMN diagram interchange (layout + shapes + edges).
-
-    Sizes every node, computes a layered layout, then emits a BPMNShape per node
-    and a BPMNEdge per flow. Edges are routed orthogonally: vertical segments
-    ("risers") live in the gaps between columns, never inside one, and each riser
-    in a gap gets its own channel so arrows do not run on top of one another.
-    Edges that skip layers are threaded through reserved dummy lanes (see
-    :func:`_sugiyama_layout`) so they do not cross the nodes in between.
-    ``verify`` upstream guarantees every node and flow endpoint exists, so
-    positions are looked up directly.
-    """
-    bpmn_di = ET.SubElement(
-        definitions, f"{{{_NS['bpmndi']}}}BPMNDiagram", attrib={"id": "BPMNDiagram_1"}
-    )
-    bpmn_plane = ET.SubElement(
-        bpmn_di,
-        f"{{{_NS['bpmndi']}}}BPMNPlane",
-        attrib={"id": "BPMNPlane_1", "bpmnElement": "Process_1"},
-    )
-
-    sizes: dict[str, dict] = {}
-    for event in model["events"]:
-        sizes[event["id"]] = {"w": _BPMN_EVENT_W, "h": _BPMN_EVENT_H}
-    for task in model["tasks"]:
-        sizes[task["id"]] = {"w": _BPMN_TASK_W, "h": _BPMN_TASK_H}
-    for gateway in model["gateways"]:
-        sizes[gateway["id"]] = {"w": _BPMN_GATEWAY_W, "h": _BPMN_GATEWAY_H}
-
-    positions, ctx = _sugiyama_layout(
-        sizes, model["flows"], h_gap=180, v_gap=90, x_offset=50, y_offset=50
-    )
-
-    for element in model["tasks"] + model["events"] + model["gateways"]:
-        bpmn_shape = ET.SubElement(
-            bpmn_plane,
-            f"{{{_NS['bpmndi']}}}BPMNShape",
-            attrib={"id": f"{element['id']}_di", "bpmnElement": element["id"]},
-        )
-        pos = positions[element["id"]]
-        ET.SubElement(
-            bpmn_shape,
-            f"{{{_NS['dc']}}}Bounds",
-            attrib={
-                "x": str(pos["x"]),
-                "y": str(pos["y"]),
-                "width": str(pos["w"]),
-                "height": str(pos["h"]),
-            },
-        )
-
-    if not model["flows"]:
-        return
-
-    routes = _route_edges(positions, ctx, model["flows"], "full_ortho")
-    for idx, flow in enumerate(model["flows"]):
-        bpmn_edge = ET.SubElement(
-            bpmn_plane,
-            f"{{{_NS['bpmndi']}}}BPMNEdge",
-            attrib={"id": f"{flow['id']}_di", "bpmnElement": flow["id"]},
-        )
-        prev = None
-        for px, py in routes[idx]:
-            point = (str(int(round(px))), str(int(round(py))))
-            if point == prev:
-                continue
-            ET.SubElement(
-                bpmn_edge,
-                f"{{{_NS['di']}}}waypoint",
-                attrib={"x": point[0], "y": point[1]},
-            )
-            prev = point
-
-
-def json_to_bpmn(model, include_layout=True):
-    """Convert a validated logical process model into BPMN 2.0 XML.
-
-    Builds the semantic process and serializes it. When ``include_layout`` is
-    true it also draws the diagram (shapes and waypoints). The PNML path passes
-    ``include_layout=False``: the transformer ignores BPMN layout and we lay the
-    PNML out separately, so computing a BPMN layout there would be wasted work.
-    """
-    logger.info(
-        "Converting model to BPMN",
-        extra={k: len(model[k]) for k in ("events", "tasks", "gateways", "flows")},
-    )
-    definitions = _build_semantic_process(model)
-    if include_layout:
-        _add_diagram(definitions, model)
-
-    tree = ET.ElementTree(definitions)
-    ET.indent(tree, space="  ", level=0)
-    return ET.tostring(definitions, encoding="utf-8", xml_declaration=True).decode(
-        "utf-8"
-    )
